extract the attrs code and better document meaning
