+++ crates/formality-macros/src/term.rs
@@ -1,8 +1,9 @@
 use proc_macro2::TokenStream;
 use quote::quote;
-use syn::{Attribute, DeriveInput};
+use syn::DeriveInput;
 
 use crate::{
+    attrs::remove_formality_attributes,
     cast::{downcast_impls, upcast_impls},
     debug::derive_debug_with_spec,
     fold::derive_fold,
@@ -35,18 +36,6 @@ pub fn term(spec: Option<FormalitySpec>, mut input: DeriveInput) -> syn::Result<
     })
 }
 
-fn remove_formality_attributes(input: &mut DeriveInput) {
-    if let syn::Data::Enum(v) = &mut input.data {
-        for variant in &mut v.variants {
-            variant.attrs.retain(|attr| {
-                !attr.path().is_ident("grammar")
-                    && !attr.path().is_ident("cast")
-                    && !attr.path().is_ident("variable")
-            });
-        }
-    }
-}
-
 fn derive_term(mut s: synstructure::Structure) -> TokenStream {
     s.underscore_const(true);
     s.bind_with(|_| synstructure::BindStyle::Move);
@@ -58,8 +47,4 @@ fn derive_term(mut s: synstructure::Structure) -> TokenStream {
         gen impl CoreTerm<crate::FormalityLang> for @Self {
         }
     })
-}
-
-pub(crate) fn has_variable_attr(attrs: &[Attribute]) -> bool {
-    attrs.iter().any(|a| a.path().is_ident("variable"))
 }
